✨ Add loading state to the dialog
The changes in this commit add a loading state to the project edit dialog. This is done by adding a new `loading` prop to the `ProjectEditDialog` component and using it to display a loading indicator on the save button.

Additionally, the `ProjectCard` component has been updated to set the `loading` prop on the `ProjectEditDialog` component when the `onEditSubmit` function is called. This ensures that the loading indicator is displayed while the project update is in progress.